chore: make hapi instrumentation support @hapi/hapi rename
