+++ .travis.yml
@@ -100,7 +100,7 @@ jobs:
     -
       node_js: '12'
       if: type IN (cron, pull_request) AND NOT branch =~ ^greenkeeper/.*
-      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,express,express-queue
+      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,@hapi/hapi,express,express-queue
       script: tav --quiet
     -
       node_js: '12'
@@ -127,7 +127,7 @@ jobs:
     -
       node_js: '11'
       if: type IN (cron, pull_request) AND NOT branch =~ ^greenkeeper/.*
-      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,express,express-queue
+      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,@hapi/hapi,express,express-queue
       script: tav --quiet
     -
       node_js: '11'
@@ -154,7 +154,7 @@ jobs:
     -
       node_js: '10'
       if: type IN (cron, pull_request) AND NOT branch =~ ^greenkeeper/.*
-      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,express,express-queue
+      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,@hapi/hapi,express,express-queue
       script: tav --quiet
     -
       node_js: '10'
@@ -181,7 +181,7 @@ jobs:
     -
       node_js: '8'
       if: type IN (cron, pull_request) AND NOT branch =~ ^greenkeeper/.*
-      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,express,express-queue
+      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,@hapi/hapi,express,express-queue
       script: tav --quiet
     -
       node_js: '8'
@@ -208,7 +208,7 @@ jobs:
     -
       node_js: '6'
       if: type IN (cron, pull_request) AND NOT branch =~ ^greenkeeper/.*
-      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,express,express-queue
+      env: TAV=ws,graphql,express-graphql,elasticsearch,hapi,@hapi/hapi,express,express-queue
       script: tav --quiet
     -
       node_js: '6'
