+++ docs/agent-api.asciidoc
@@ -690,12 +690,12 @@ Optionally, a type may also be provided to group lambdas together. By default,
 Read more lambda support in the <<lambda,Lambda>> article.
 
 [[apm-add-patch]]
-==== `apm.addPatch(name, handler)`
+==== `apm.addPatch(modules, handler)`
 
 [small]#Added in: v2.7.0#
 
-* `name` +{type-string}+
-Name of module to apply the patch to, when required.
+* `modules` +{type-string}+ | +{type-string-array}+
+Name of module(s) to apply the patch to, when required.
 * `handler` +{type-function}+ | +{type-string}+
 Must be a patch function or a path to a module exporting a patch function
 ** `exports` +{type-object}+ The original export object of the module
@@ -726,11 +726,26 @@ apm.addPatch('timers', (exports, agent, { version, enabled }) => {
 
 // or ...
 
+apm.addPatch(['hapi', '@hapi/hapi'], (exports, agent, { version, enabled }) => {
+  const setTimeout = exports.setTimeout
+  exports.setTimeout = (fn, ms) => {
+    const span = agent.createSpan('set-timeout')
+    return setTimeout(() => {
+      span.end()
+      fn()
+    }, ms)
+  }
+
+  return exports
+})
+
+// or ...
+
 apm.addPatch('timers', './timer-patch')
 ----
 
 [[apm-remove-patch]]
-==== `apm.removePatch(name, handler)`
+==== `apm.removePatch(modules, handler)`
 
 [small]#Added in: v2.7.0#
 
@@ -741,6 +756,10 @@ To _disable_ instrumentation while keeping context propagation support, see <<di
 [source,js]
 ----
 apm.removePatch('timers', './timers-patch')
+
+// or ...
+
+apm.removePatch(['timers'], './timers-patch')
 
 // or ...
