+++ test/instrumentation/modules/hapi/basic-legacy-path.js
@@ -1,0 +1,3 @@
+'use strict'
+
+require('./shared')('hapi')
